chore(deps): handle css files through job-ripper
