+++ package.json
@@ -71,6 +71,7 @@
     "markdown-it": "^14.2.0",
     "nodemon": "^3.1.14",
     "postcss": "^8.5.15",
+    "postcss-import": "^16.1.1",
     "postcss-less": "^6.0.0",
     "rimraf": "^6.1.3",
     "smoothscroll-polyfill": "^0.4.4",
